feat: splittable POS items for KG/L/M units - decimal quantities
- ProductGrid: Clicking KG/L/M items opens quantity input popup (step 0.01)
- Cart: KG/L/M items show direct decimal input field with unit label
- Cart: Quantity items keep +/- button controls
- ProductForm: stock_quantity input now allows decimals (step 0.01)
- Backend: stock_quantity validation changed from isInt to isFloat
- Schema: stock_quantity changed from INTEGER to NUMERIC
- Migration: server/migrations/003_stock_quantity_to_numeric.sql

+++ server/src/db/seed.js
@@ -55,7 +55,7 @@ CREATE TABLE IF NOT EXISTS products (
   category_id BIGINT REFERENCES categories(id) ON DELETE SET NULL,
   price NUMERIC NOT NULL DEFAULT 0,
   cost_price NUMERIC DEFAULT 0,
-  stock_quantity INTEGER DEFAULT 0,
+  stock_quantity NUMERIC DEFAULT 0,
   low_stock_threshold INTEGER DEFAULT 10,
   is_refundable BOOLEAN DEFAULT true,
   unit_of_measure TEXT DEFAULT 'quantity',
